moved pages/pharmacy/stock components into _component folder
