+++ pages/pharmacy/stock/_components/index.js
@@ -1,0 +1,3 @@
+import { MedicationModal } from "./MedicationModal";
+
+export { MedicationModal };
